staslib: break stas.py into smaller modules
The following modules were created from stas.py.
- conf.py
- gutil.py
- log.py
- trid.py
- udev.py
