+++ stacd.py
@@ -99,23 +99,20 @@ stas.check_if_allowed_to_continue()
 # pylint: disable=wrong-import-order
 import json
 import pathlib
-import logging
 import systemd.daemon
 import dasbus.error
 import dasbus.client.observer
 import dasbus.client.proxy
 from libnvme import nvme
 from gi.repository import GLib
+from staslib import conf, log, gutil, trid, udev  # pylint: disable=ungrouped-imports
 
-stas.LOG.addHandler(stas.get_log_handler(ARGS.syslog, defs.STACD_PROCNAME))
-stas.LOG.setLevel(logging.INFO if ARGS.syslog else logging.DEBUG)
-stas.CNF.conf_file = ARGS.conf_file
-stas.trace_control(ARGS.tron or stas.CNF.tron)
+log.init(ARGS.syslog)
+conf.PROCESS.conf_file = ARGS.conf_file
+stas.trace_control(ARGS.tron or conf.PROCESS.tron)
 
-NVME_ROOT = nvme.root()  # Singleton
-NVME_ROOT.log_level("debug" if (ARGS.tron or stas.CNF.tron) else "err")
-NVME_HOST = nvme.host(NVME_ROOT, stas.SYS_CNF.hostnqn, stas.SYS_CNF.hostid, stas.SYS_CNF.hostsymname)  # Singleton
-
+NVME_HOST = nvme.host(stas.NVME_ROOT, conf.SYSTEM.hostnqn, conf.SYSTEM.hostid, conf.SYSTEM.hostsymname)  # Singleton
+UDEV = udev.Udev()
 UDEV_RULE_SUPPRESS = pathlib.Path('/run/udev/rules.d', '70-nvmf-autoconnect.rules')
 
 
@@ -137,29 +134,24 @@ def udev_rule_ctrl(enable):
             UDEV_RULE_SUPPRESS.symlink_to('/dev/null')
 
 
-def set_loglevel(tron):  # pylint: disable=missing-function-docstring
-    stas.trace_control(tron)
-    NVME_ROOT.log_level("debug" if tron else "err")
-
-
 # ******************************************************************************
 class Ioc(stas.Controller):
     '''@brief This object establishes a connection to one I/O Controller.'''
 
-    def __init__(self, tid: stas.TransportId):
-        super().__init__(NVME_ROOT, NVME_HOST, tid)
+    def __init__(self, tid: trid.TID):
+        super().__init__(stas.NVME_ROOT, NVME_HOST, tid)
 
-    def _on_udev_remove(self, udev):
+    def _on_udev_remove(self, udev_obj):
         '''Called when the associated nvme device (/dev/nvmeX) is removed
         from the system.
         '''
-        super()._on_udev_remove(udev)
+        super()._on_udev_remove(udev_obj)
 
         # Defer removal of this object to the next main loop's idle period.
         GLib.idle_add(STAC.remove_controller, self)
 
     def _find_existing_connection(self):
-        return stas.UDEV.find_nvme_ioc_device(self.tid)
+        return UDEV.find_nvme_ioc_device(self.tid)
 
 
 # ******************************************************************************
@@ -184,12 +176,12 @@ class Stac(stas.Service):
         @tron.setter
         def tron(self, value):  # pylint: disable=no-self-use
             '''@brief Set Trace ON property'''
-            set_loglevel(value)
+            stas.trace_control(value)
 
         @property
         def log_level(self) -> str:
             '''@brief Get Log Level property'''
-            return stas.log_level()
+            return log.level()
 
         def process_info(self) -> str:
             '''@brief Get status info (for debug)
@@ -226,7 +218,7 @@ class Stac(stas.Service):
         # to the system. The following timer acts as a "soak period". Changes
         # will be applied by calling self._on_config_ctrls() at the end of
         # the soak period.
-        self._cfg_soak_tmr = stas.GTimer(Stac.CONF_STABILITY_SOAK_TIME_SEC, self._on_config_ctrls)
+        self._cfg_soak_tmr = gutil.GTimer(Stac.CONF_STABILITY_SOAK_TIME_SEC, self._on_config_ctrls)
         self._cfg_soak_tmr.start()
 
         self._config_connections_audit()
@@ -242,14 +234,14 @@ class Stac(stas.Service):
         self._staf_watcher.connect_once_available()
 
         # Suppress udev rule to auto-connect when AEN is received.
-        udev_rule_ctrl(stas.CNF.udev_rule_enabled)
+        udev_rule_ctrl(conf.PROCESS.udev_rule_enabled)
 
     def _release_resources(self):
-        stas.LOG.debug('Stac._release_resources()')
+        log.LOG.debug('Stac._release_resources()')
 
         udev_rule_ctrl(True)
 
-        stas.UDEV.unregister_for_action_events('add')
+        UDEV.unregister_for_action_events('add')
 
         self._destroy_staf_comlink(self._staf_watcher)
         if self._staf_watcher is not None:
@@ -268,7 +260,7 @@ class Stac(stas.Service):
         stacd.conf. stacd will audit connections if "sticky-connections=disabled".
         stacd will delete any connection that is not supposed to exist.
         '''
-        stas.LOG.debug('Stac._audit_connections()          - tids = %s', tids)
+        log.LOG.debug('Stac._audit_connections()          - tids = %s', tids)
         num_controllers = len(self._controllers)
         for tid in tids:
             if tid not in self._controllers:
@@ -277,14 +269,14 @@ class Stac(stas.Service):
         if num_controllers != len(self._controllers):
             self._cfg_soak_tmr.start(Stac.CONF_STABILITY_SOAK_TIME_SEC)
 
-    def _on_add_event(self, udev):
+    def _on_add_event(self, udev_obj):
         '''@brief This function is called when a "add" event is received from
         the kernel for an NVMe device. This is used to trigger an audit and make
         sure that the connection to an I/O controller is allowed.
         '''
-        if stas.UDEV.is_ioc_device(udev):
-            tid = stas.UDEV.get_tid(udev)
-            stas.LOG.debug('Stac._on_add_event()               - tid=%s | %s', tid, udev.sys_name)
+        if UDEV.is_ioc_device(udev_obj):
+            tid = UDEV.get_tid(udev_obj)
+            log.LOG.debug('Stac._on_add_event()               - tid=%s | %s', tid, udev_obj.sys_name)
             self._audit_connections([tid])
 
     def _config_connections_audit(self):
@@ -292,12 +284,12 @@ class Stac(stas.Service):
         whether audits should be performed. Audits are enabled when
         "sticky_connections" is disabled.
         '''
-        if not stas.CNF.sticky_connections:
-            if stas.UDEV.get_registered_action_cback('add') is None:
-                stas.UDEV.register_for_action_events('add', self._on_add_event)
-                self._audit_connections(stas.UDEV.get_nvme_ioc_tids())
+        if not conf.PROCESS.sticky_connections:
+            if UDEV.get_registered_action_cback('add') is None:
+                UDEV.register_for_action_events('add', self._on_add_event)
+                self._audit_connections(UDEV.get_nvme_ioc_tids())
         else:
-            stas.UDEV.unregister_for_action_events('add')
+            UDEV.unregister_for_action_events('add')
 
     def _keep_connections_on_exit(self):
         '''@brief Determine whether connections should remain when the
@@ -310,11 +302,11 @@ class Stac(stas.Service):
         signal, which can be sent with "systemctl reload stacd".
         '''
         systemd.daemon.notify('RELOADING=1')
-        stas.CNF.reload()
-        set_loglevel(stas.CNF.tron)
+        conf.PROCESS.reload()
+        stas.trace_control(conf.PROCESS.tron)
         self._config_connections_audit()
         self._cfg_soak_tmr.start(Stac.CONF_STABILITY_SOAK_TIME_SEC)
-        udev_rule_ctrl(stas.CNF.udev_rule_enabled)
+        udev_rule_ctrl(conf.PROCESS.udev_rule_enabled)
         systemd.daemon.notify('READY=1')
         return GLib.SOURCE_CONTINUE
 
@@ -322,7 +314,7 @@ class Stac(stas.Service):
         configured_ctrl_list = [
             ctrl_dict for ctrl_dict in configured_ctrl_list if 'traddr' in ctrl_dict and 'subsysnqn' in ctrl_dict
         ]
-        stas.LOG.debug('Stac._config_ctrls_finish()        - configured_ctrl_list = %s', configured_ctrl_list)
+        log.LOG.debug('Stac._config_ctrls_finish()        - configured_ctrl_list = %s', configured_ctrl_list)
 
         discovered_ctrl_list = list()
         if self._staf:
@@ -332,23 +324,23 @@ class Stac(stas.Service):
                 for dlpe in staf_data['log-pages']:
                     if dlpe.get('subtype') == 'nvme':  # eliminate discovery controllers
                         discovered_ctrl_list.append(stas.cid_from_dlpe(dlpe, host_traddr, host_iface))
-        stas.LOG.debug('Stac._config_ctrls_finish()        - discovered_ctrl_list = %s', discovered_ctrl_list)
+        log.LOG.debug('Stac._config_ctrls_finish()        - discovered_ctrl_list = %s', discovered_ctrl_list)
 
         controllers = stas.remove_blacklisted(configured_ctrl_list + discovered_ctrl_list)
         controllers = stas.remove_invalid_addresses(controllers)
 
-        new_controller_ids = {stas.TransportId(controller) for controller in controllers}
+        new_controller_ids = {trid.TID(controller) for controller in controllers}
         cur_controller_ids = set(self._controllers.keys())
         controllers_to_add = new_controller_ids - cur_controller_ids
         controllers_to_del = cur_controller_ids - new_controller_ids
 
-        stas.LOG.debug('Stac._config_ctrls_finish()        - controllers_to_add   = %s', list(controllers_to_add))
-        stas.LOG.debug('Stac._config_ctrls_finish()        - controllers_to_del   = %s', list(controllers_to_del))
+        log.LOG.debug('Stac._config_ctrls_finish()        - controllers_to_add   = %s', list(controllers_to_add))
+        log.LOG.debug('Stac._config_ctrls_finish()        - controllers_to_del   = %s', list(controllers_to_del))
 
         for tid in controllers_to_del:
             controller = self._controllers.pop(tid, None)
             if controller is not None:
-                controller.disconnect(self.remove_controller, stas.CNF.sticky_connections)
+                controller.disconnect(self.remove_controller, conf.PROCESS.sticky_connections)
 
         for tid in controllers_to_add:
             self._controllers[tid] = Ioc(tid)
@@ -362,9 +354,9 @@ class Stac(stas.Service):
 
             # Make sure timer is set back to its normal value.
             self._cfg_soak_tmr.set_timeout(Stac.CONF_STABILITY_SOAK_TIME_SEC)
-            stas.LOG.debug('Stac._connect_to_staf()            - Connected to staf')
+            log.LOG.debug('Stac._connect_to_staf()            - Connected to staf')
         except dasbus.error.DBusError:
-            stas.LOG.error('Failed to connect to staf')
+            log.LOG.error('Failed to connect to staf')
 
     def _destroy_staf_comlink(self, watcher):  # pylint: disable=unused-argument
         if self._staf:
@@ -384,12 +376,12 @@ class Stac(stas.Service):
         if self._cfg_soak_tmr:
             self._cfg_soak_tmr.set_timeout(Stac.CONF_STABILITY_LONG_SOAK_TIME_SEC)
 
-        stas.LOG.debug('Stac._disconnect_from_staf()       - Disconnected from staf')
+        log.LOG.debug('Stac._disconnect_from_staf()       - Disconnected from staf')
 
     def _log_pages_changed(  # pylint: disable=too-many-arguments
         self, transport, traddr, trsvcid, host_traddr, host_iface, subsysnqn, device
     ):
-        stas.LOG.debug(
+        log.LOG.debug(
             'Stac._log_pages_changed()          - transport=%s, traddr=%s, trsvcid=%s, host_traddr=%s, host_iface=%s, subsysnqn=%s, device=%s',
             transport,
             traddr,
@@ -413,7 +405,11 @@ if __name__ == '__main__':
     STAC = Stac()
     STAC.run()
 
+    UDEV = None
     STAC = None
     ARGS = None
+
     stas.clean()
-    logging.shutdown()
+    conf.clean()
+    udev.clean()
+    log.clean()
