+++ staslib/avahi.py
@@ -17,7 +17,7 @@ import dasbus.connection
 import dasbus.client.proxy
 import dasbus.client.observer
 from gi.repository import GLib
-from staslib import stas, defs
+from staslib import defs, log, conf, gutil
 
 
 def txt2dict(txt: list):
@@ -100,7 +100,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
         # access it over D-Bus to wake it up. The following timer is used to
         # periodically query the avahi-daemon until we successfully establish
         # first contact.
-        self._kick_avahi_tmr = stas.GTimer(60, self._on_kick_avahi)
+        self._kick_avahi_tmr = gutil.GTimer(60, self._on_kick_avahi)
 
         # Subscribe for Avahi signals (i.e. events). This must be done before
         # any Browser or Resolver is created to avoid race conditions and
@@ -144,7 +144,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
 
     def kill(self):
         '''@brief Clean up object'''
-        stas.LOG.debug('Avahi.kill()')
+        log.LOG.debug('Avahi.kill()')
 
         self._kick_avahi_tmr.kill()
         self._kick_avahi_tmr = None
@@ -223,7 +223,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
         self._kick_avahi_tmr.clear()
 
     def _disconnect(self):
-        stas.LOG.debug('Avahi._disconnect()')
+        log.LOG.debug('Avahi._disconnect()')
         for service in self._services.values():
             resolver = service.pop('resolver', None)
             if resolver is not None:
@@ -231,7 +231,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
                     resolver.Free()
                     dasbus.client.proxy.disconnect_proxy(resolver)
                 except (AttributeError, dasbus.error.DBusError) as ex:
-                    stas.LOG.debug('Avahi._disconnect()                - Failed to Free() resolver. %s', ex)
+                    log.LOG.debug('Avahi._disconnect()                - Failed to Free() resolver. %s', ex)
 
         self._services = dict()
 
@@ -240,7 +240,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
                 browser.Free()
                 dasbus.client.proxy.disconnect_proxy(browser)
             except (AttributeError, dasbus.error.DBusError) as ex:
-                stas.LOG.debug('Avahi._disconnect()                - Failed to Free() browser. %s', ex)
+                log.LOG.debug('Avahi._disconnect()                - Failed to Free() browser. %s', ex)
 
         self._service_browsers = dict()
 
@@ -256,14 +256,14 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
 
     def _avahi_available(self, _avahi_watcher):
         '''@brief Hook up DBus signal handlers for signals from stafd.'''
-        stas.LOG.info('avahi-daemon service available, zeroconf supported.')
+        log.LOG.info('avahi-daemon service available, zeroconf supported.')
         success = self._configure_browsers()
         if not success:
             self._kick_avahi_tmr.start()
 
     def _avahi_unavailable(self, _avahi_watcher):
         self._disconnect()
-        stas.LOG.warning('avahi-daemon not available, zeroconf not supported.')
+        log.LOG.warning('avahi-daemon not available, zeroconf not supported.')
         self._kick_avahi_tmr.start()
 
     def _configure_browsers(self):
@@ -271,8 +271,8 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
         stypes_to_add = self._stypes - stypes_cur
         stypes_to_rm = stypes_cur - self._stypes
 
-        stas.LOG.debug('Avahi._configure_browsers()        - stypes_to_rm  = %s', list(stypes_to_rm))
-        stas.LOG.debug('Avahi._configure_browsers()        - stypes_to_add = %s', list(stypes_to_add))
+        log.LOG.debug('Avahi._configure_browsers()        - stypes_to_rm  = %s', list(stypes_to_rm))
+        log.LOG.debug('Avahi._configure_browsers()        - stypes_to_add = %s', list(stypes_to_add))
 
         for stype_to_rm in stypes_to_rm:
             browser = self._service_browsers.pop(stype_to_rm, None)
@@ -281,7 +281,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
                     browser.Free()
                     dasbus.client.proxy.disconnect_proxy(browser)
                 except (AttributeError, dasbus.error.DBusError) as ex:
-                    stas.LOG.debug('Avahi._configure_browsers()        - Failed to Free() browser. %s', ex)
+                    log.LOG.debug('Avahi._configure_browsers()        - Failed to Free() browser. %s', ex)
 
             # Find the cached services corresponding to stype_to_rm and remove them
             services_to_rm = [service for service in self._services if service[3] == stype_to_rm]
@@ -292,7 +292,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
                         resolver.Free()
                         dasbus.client.proxy.disconnect_proxy(resolver)
                     except (AttributeError, dasbus.error.DBusError) as ex:
-                        stas.LOG.debug('Avahi._configure_browsers()        - Failed to Free() resolver. %s', ex)
+                        log.LOG.debug('Avahi._configure_browsers()        - Failed to Free() resolver. %s', ex)
 
         for stype in stypes_to_add:
             try:
@@ -301,8 +301,8 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
                 )
                 self._service_browsers[stype] = self._sysbus.get_proxy(Avahi.DBUS_NAME, obj_path)
             except dasbus.error.DBusError as ex:
-                stas.LOG.debug('Avahi._configure_browsers()        - Failed to contact avahi-daemon. %s', ex)
-                stas.LOG.warning('avahi-daemon not available, operating w/o mDNS discovery.')
+                log.LOG.debug('Avahi._configure_browsers()        - Failed to contact avahi-daemon. %s', ex)
+                log.LOG.warning('avahi-daemon not available, operating w/o mDNS discovery.')
                 return False
 
         return True
@@ -318,7 +318,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
         *_user_data
     ):
         (interface, protocol, name, stype, domain, flags) = args
-        stas.LOG.debug(
+        log.LOG.debug(
             'Avahi._service_discovered()        - interface=%s (%s), protocol=%s, stype=%s, domain=%s, flags=%s %-14s name=%s',
             interface,
             socket.if_indextoname(interface),
@@ -341,7 +341,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
                     'data': {},
                 }
             except dasbus.error.DBusError as ex:
-                stas.LOG.warning('Failed to create resolver: "%s", "%s", "%s". %s', interface, name, stype, ex)
+                log.LOG.warning('Failed to create resolver: "%s", "%s", "%s". %s', interface, name, stype, ex)
 
     def _service_removed(
         self,
@@ -354,7 +354,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
         *_user_data
     ):
         (interface, protocol, name, stype, domain, flags) = args
-        stas.LOG.debug(
+        log.LOG.debug(
             'Avahi._service_removed()           - interface=%s (%s), protocol=%s, stype=%s, domain=%s, flags=%s %-14s name=%s',
             interface,
             socket.if_indextoname(interface),
@@ -373,7 +373,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
                 resolver.Free()
                 dasbus.client.proxy.disconnect_proxy(resolver)
             except (AttributeError, dasbus.error.DBusError) as ex:
-                stas.LOG.debug('Avahi._service_removed()           - Failed to Free() resolver. %s', ex)
+                log.LOG.debug('Avahi._service_removed()           - Failed to Free() resolver. %s', ex)
 
         self._change_cb()
 
@@ -389,7 +389,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
     ):
         (interface, protocol, name, stype, domain, host, aprotocol, address, port, txt, flags) = args
         txt = txt2dict(txt)
-        stas.LOG.debug(
+        log.LOG.debug(
             'Avahi._service_identified()        - interface=%s (%s), protocol=%s, stype=%s, domain=%s, flags=%s %-14s name=%s, host=%s, aprotocol=%s, address=%s, port=%s, txt=%s',
             interface,
             socket.if_indextoname(interface),
@@ -414,7 +414,7 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
                 'trsvcid':    str(port),
                 'host-iface': socket.if_indextoname(interface),
                 'subsysnqn':  txt.get('NQN', defs.WELL_KNOWN_DISC_NQN)
-                              if stas.NvmeOptions().discovery_supp
+                              if conf.NvmeOptions().discovery_supp
                               else defs.WELL_KNOWN_DISC_NQN,
             }
         self._change_cb()
@@ -432,4 +432,4 @@ class Avahi:  # pylint: disable=too-many-instance-attributes
         (error,) = args
         if 'ServiceResolver' not in interface_name or 'TimeoutError' not in error:
             # ServiceResolver may fire a timeout event after being Free'd(). This seems to be normal.
-            stas.LOG.error('Avahi._failure_handler()    - name=%s, error=%s', interface_name, error)
+            log.LOG.error('Avahi._failure_handler()    - name=%s, error=%s', interface_name, error)
