+++ usr/lib/systemd/system/stafd.in.service
@@ -22,8 +22,6 @@ Type=dbus
 BusName=@STAFD_DBUS_NAME@
 SyslogIdentifier=@STAFD_PROCNAME@
 
-Environment=XDG_CACHE_HOME=/var/cache/@STAFD_PROCNAME@
-
 ExecStart=/usr/bin/python3 -u @STAFD_EXECUTABLE@ --syslog
 ExecReload=/bin/kill -HUP $MAINPID
